# core/src/main/java/ru/stand/contest/restapi_yandex/controller/NodesApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.0.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package ru.stand.contest.restapi_yandex.controller;

import ru.stand.contest.restapi_yandex.model.Error;
import ru.stand.contest.restapi_yandex.model.SystemItem;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;

import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2022-09-09T13:45:12.975114900+03:00[Europe/Moscow]")
@Validated
@Tag(name = "nodes", description = "the nodes API")
public interface NodesApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    ResponseEntity<SystemItem> getSystemItem(String id);

    /**
     * GET /nodes/{id}
     * Получить информацию об элементе по идентификатору. При получении информации о папке также предоставляется информация о её дочерних элементах.  - для пустой папки поле children равно пустому массиву, а для файла равно null - размер папки - это суммарный размер всех её элементов. Если папка не содержит элементов, то размер равен 0. При обновлении размера элемента, суммарный размер папки, которая содержит этот элемент, тоже обновляется. 
     *
     * @param id Идентификатор элемента (required)
     * @return Информация об элементе. (status code 200)
     *         or Невалидная схема документа или входные данные не верны. (status code 400)
     *         or Элемент не найден. (status code 404)
     */
    @Operation(
        operationId = "nodesIdGet",
        tags = { "Базовые задачи" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Информация об элементе.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SystemItem.class))
            }),
            @ApiResponse(responseCode = "400", description = "Невалидная схема документа или входные данные не верны.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "404", description = "Элемент не найден.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/nodes/{id}",
        produces = { "application/json" }
    )
    default ResponseEntity<SystemItem> nodesIdGet(
        @Parameter(name = "id", description = "Идентификатор элемента", required = true) @PathVariable("id") String id
    ) {
//        getRequest().ifPresent(request -> {
//            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
//                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
//                    String exampleString = "{ \"id\" : \"элемент_1_2\", \"url\" : null, \"type\" : \"FOLDER\", \"parentId\" : null, \"date\" : \"2022-05-28T21:12:01.000Z\", \"size\" : 12, \"children\" : [ { \"url\" : \"/file/url1\", \"id\" : \"элемент_1_3\", \"size\" : 4, \"date\" : \"2022-05-28T21:12:01.000Z\", \"type\" : \"FILE\", \"parentId\" : \"элемент_1_2\" }, { \"type\" : \"FOLDER\", \"url\" : null, \"id\" : \"элемент_1_1\", \"date\" : \"2022-05-26T21:12:01.000Z\", \"parentId\" : \"элемент_1_2\", \"size\" : 8, \"children\" : [ { \"url\" : \"/file/url2\", \"id\" : \"элемент_1_4\", \"parentId\" : \"элемент_1_1\", \"date\" : \"2022-05-26T21:12:01.000Z\", \"size\" : 8, \"type\" : \"FILE\" } ] } ] }";
//                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
//                    break;
//                }
//            }
//        });
        return getSystemItem(id);

    }

}